requirements.txt
python-dotenv
jupyter
azure-identity
openai
tiktoken
azure-search-documents==11.6.0b6
azure-ai-documentintelligence==1.0.0b2
langchain
langchain_openai
colorama
Jinja2
gradio
yfinance
chainlit==1.1.401
